// frontend/src/api/generated/models/ApartmentStatisticsDTO.ts
export interface ApartmentStatisticsDTO {
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    apartmentId: number;
    /**
     * 
     * @type {string}
     * @memberof ApartmentStatisticsDTO
     */
    apartmentName: string;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    totalRooms: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    occupiedRooms: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    freeRooms: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    occupancyRatePercent: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    activeContractsCount: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    totalContractsCount: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    activeMonthlyRent: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    averageRentPerOccupiedRoom: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    averageRentAllContracts: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    totalIncomeCollected: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    currentMonthIncomeCollected: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    totalCosts: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    currentMonthCosts: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    averageMonthlyCost: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    netResultCurrentMonth: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentStatisticsDTO
     */
    netResultAllTime: number;
}